more array, struct warns

+++ shedskin/lib/struct.cpp
@@ -302,12 +302,12 @@ void fillbuf_int(char c, __ss_int t, char order, unsigned int itemsize) {
             case 'B': *((unsigned char *)buffy) = (unsigned char)t; break;
             case 'h': *((short *)buffy) = (short)t; break;
             case 'H': *((unsigned short *)buffy) = (unsigned short)t; break;
-            case 'i': *((int *)buffy) = t; break;
-            case 'I': *((unsigned int *)buffy) = t; break;
-            case 'l': *((long *)buffy) = t; break;
-            case 'L': *((unsigned long *)buffy) = t; break;
-            case 'q': *((long long *)buffy) = t; break;
-            case 'Q': *((unsigned long long *)buffy) = t; break;
+            case 'i': *((int *)buffy) = (int)t; break;
+            case 'I': *((unsigned int *)buffy) = (unsigned int)t; break;
+            case 'l': *((long *)buffy) = (long)t; break;
+            case 'L': *((unsigned long *)buffy) = (unsigned long)t; break;
+            case 'q': *((long long *)buffy) = (long long)t; break;
+            case 'Q': *((unsigned long long *)buffy) = (unsigned long long)t; break;
 //            case 'n': *((ssize_t *)buffy) = t; break;
             case 'N': *((size_t *)buffy) = (size_t)t; break;
         }
